(#6) Ajuste geral lógica post e Home
